LPS-59517 move files so this package is exported
