+++ portal-impl/src/com/liferay/portal/action/UpdateLicenseAction.java
@@ -17,12 +17,12 @@ package com.liferay.portal.action;
 import com.liferay.portal.kernel.json.JSONArray;
 import com.liferay.portal.kernel.json.JSONFactoryUtil;
 import com.liferay.portal.kernel.json.JSONObject;
+import com.liferay.portal.kernel.license.util.LicenseManagerUtil;
 import com.liferay.portal.kernel.servlet.ServletResponseUtil;
 import com.liferay.portal.kernel.util.Constants;
 import com.liferay.portal.kernel.util.ContentTypes;
 import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.util.StringPool;
-import com.liferay.portal.license.util.LicenseManagerUtil;
 import com.liferay.portal.license.util.LicenseUtil;
 import com.liferay.portal.model.User;
 import com.liferay.portal.util.PortalUtil;
