+++ portal-service/src/com/liferay/portal/kernel/license/package.html
