+++ portal-service/src/com/liferay/portal/kernel/license/packageinfo
