+++ portal-service/src/com/liferay/portal/kernel/license/util/LicenseManager.java
@@ -12,10 +12,10 @@
  * details.
  */
 
-package com.liferay.portal.license.util;
+package com.liferay.portal.kernel.license.util;
 
 import com.liferay.portal.kernel.json.JSONObject;
-import com.liferay.portal.license.LicenseInfo;
+import com.liferay.portal.kernel.license.LicenseInfo;
 
 import java.util.List;
 import java.util.Map;
