+++ portal-service/src/com/liferay/portal/kernel/license/util/package.html
