+++ portal-service/src/com/liferay/portal/kernel/license/util/packageinfo
